+++ packages/web-components/fast-foundation/src/tabs/tabs.ts
@@ -327,7 +327,10 @@ export class Tabs extends FASTElement {
 }
 
 /**
- * @public
+ * Mark internal because exporting class and interface of the same name
+ * confuses API documenter.
+ * TODO: https://github.com/microsoft/fast-dna/issues/3317
+ * @internal
  */
 /* eslint-disable-next-line */
 export interface Tabs extends StartEnd {}