+++ sites/fast-tooling-examples/app/index.ts
@@ -14,6 +14,8 @@ import "./style.css";
 import examples from "./registry";
 import toolingGuidance from "./.tmp/tooling-guidance";
 import toolingReactGuidance from "./.tmp/tooling-react-guidance";
+/* eslint-disable-next-line @typescript-eslint/no-var-requires */
+const FASTInlineLogo = require("@microsoft/site-utilities/statics/assets/fast-inline-logo.png");
 
 // prevent tree shaking
 FASTAnchor;
@@ -23,6 +25,7 @@ FASTDesignSystemProvider;
 FASTTabs;
 FASTTab;
 FASTTabPanel;
+FASTInlineLogo;
 
 /**
  * The links to examples