+++ packages/web-components/fast-components/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components@1.2.0...@microsoft/fast-components@1.2.1) (2020-06-17)
+
+
+### Bug Fixes
+
+* inherit font size and line height in hypertext styles ([#3320](https://github.com/Microsoft/fast-dna/issues/3320)) ([98348cb](https://github.com/Microsoft/fast-dna/commit/98348cbfceb35f95cfdba4f9a1706ad985a60540))
+* remove hypertext styles from button ([#3319](https://github.com/Microsoft/fast-dna/issues/3319)) ([788e299](https://github.com/Microsoft/fast-dna/commit/788e299ffed8fe965212079a01dd53d42a80b894))
+
+
+
+
+
 # [1.2.0](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components@1.1.2...@microsoft/fast-components@1.2.0) (2020-06-15)